Reworked, and expansion of outputs
This now takes a slightly more structured view on things, loading the
data in one step, then output/uploading it, and and plays around with
export formats. Now supports PDF, HTML and iCAL export

Still more work to do to, and needs a major structural rejig to move
filter logic to export step, not import step.